requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
idna==2.10
itsdangerous==1.1.0
MarkupSafe==1.1.1
multidict==5.1.0
mysqlclient==2.0.3
numpy==1.20.1
pandas==1.2.1
pydantic==1.7.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.3.23
starlette==0.13.6
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.3
websocket-client==0.58.0
Werkzeug==1.0.1
yapf==0.30.0
yarl==1.6.3
